requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
python-dotenv
uuid
bcrypt
requests
feedparser
PyMuPDF
sentence-transformers
scikit-learn
numpy
faiss-cpu
langchain
langchain-community
langchain-openai
scholarly
sqlalchemy
Werkzeug
